+++ src/amqp/tests/KeepaliveIOTest.php
@@ -14,12 +14,17 @@ namespace HyperfTest\Amqp;
 
 use Hyperf\Amqp\Connection\KeepaliveIO;
 use Hyperf\Amqp\Connection\Socket;
+use Hyperf\Contract\StdoutLoggerInterface;
 use Hyperf\Utils\Context;
+use Hyperf\Utils\Str;
 use HyperfTest\Amqp\Stub\ContainerStub;
 use HyperfTest\Amqp\Stub\SocketStub;
+use HyperfTest\Amqp\Stub\SocketWithoutIOStub;
 use Mockery;
+use PhpAmqpLib\Exception\AMQPRuntimeException;
 use PhpAmqpLib\Wire\AMQPWriter;
 use PHPUnit\Framework\TestCase;
+use Swoole\Coroutine;
 use Swoole\Timer;
 
 /**
@@ -44,7 +49,7 @@ class KeepaliveIOTest extends TestCase
 
         $container = ContainerStub::getHyperfContainer();
         $container->shouldReceive('make')->with(Socket::class, Mockery::any())->andReturnUsing(function ($_, $args) use ($host, $port, $timeout, $heartbeat) {
-            $this->assertEquals([$host, $port, $timeout, $heartbeat], $args);
+            $this->assertEquals(['host' => $host, 'port' => $port, 'timeout' => $timeout, 'heartbeat' => $heartbeat], $args);
             return Mockery::mock(Socket::class);
         });
 
@@ -61,7 +66,7 @@ class KeepaliveIOTest extends TestCase
 
         $container = ContainerStub::getHyperfContainer();
         $container->shouldReceive('make')->with(Socket::class, Mockery::any())->andReturnUsing(function ($_, $args) {
-            return new SocketStub(...$args);
+            return new SocketStub(...array_values($args));
         });
 
         $io = new KeepaliveIO($host, $port, $timeout, 6, null, true, $heartbeat);
@@ -69,6 +74,27 @@ class KeepaliveIOTest extends TestCase
 
         $res = $io->read(10);
         $this->assertTrue(strlen($res) === 10);
+    }
+
+    public function testKeepalivePopFailed()
+    {
+        $host = '127.0.0.1';
+        $port = 5672;
+        $timeout = 5;
+        $heartbeat = 1;
+
+        $container = ContainerStub::getHyperfContainer();
+        $container->shouldReceive('make')->with(Socket::class, Mockery::any())->andReturnUsing(function ($_, $args) {
+            return new SocketWithoutIOStub(true, ...array_values($args));
+        });
+
+        $io = new KeepaliveIO($host, $port, $timeout, 6, null, true, $heartbeat);
+
+        $io->connect();
+
+        $this->expectException(AMQPRuntimeException::class);
+        $this->expectExceptionMessageRegExp('/^Socket of keepaliveIO is exhausted\. Cannot establish new socket before wait_timeout\.$/');
+        $io->read(10);
     }
 
     public function testKeepaliveHeartbeatTimer()
@@ -80,7 +106,7 @@ class KeepaliveIOTest extends TestCase
 
         $container = ContainerStub::getHyperfContainer();
         $container->shouldReceive('make')->with(Socket::class, Mockery::any())->andReturnUsing(function ($_, $args) {
-            return new SocketStub(...$args);
+            return new SocketStub(...array_values($args));
         });
 
         $io = new KeepaliveIO($host, $port, $timeout, 6, null, true, $heartbeat);
@@ -97,5 +123,59 @@ class KeepaliveIOTest extends TestCase
         $data = $pkt->getvalue();
 
         $this->assertSame($data, Context::get('test.amqp.send.data'));
+    }
+
+    public function testKeepaliveReconnect()
+    {
+        $host = '127.0.0.1';
+        $port = 5672;
+        $timeout = 5;
+        $heartbeat = 1;
+
+        $sock = new SocketWithoutIOStub(false, $host, $port, $timeout, $heartbeat);
+        $container = ContainerStub::getHyperfContainer();
+        $container->shouldReceive('make')->with(Socket::class, Mockery::any())->andReturnUsing(function ($_, $args) use ($sock) {
+            return $sock;
+        });
+
+        $io = new KeepaliveIO($host, $port, $timeout, 6, null, true, $heartbeat);
+
+        $io->connect();
+
+        try {
+            $io->read(10);
+        } catch (\Throwable $throwable) {
+            $this->assertSame(2, $sock->getConnectCount());
+        }
+    }
+
+    public function testKeepaliveIOHeartbeat()
+    {
+        $host = '127.0.0.1';
+        $port = 5672;
+        $timeout = 5;
+        $heartbeat = 1;
+
+        $sock = new SocketWithoutIOStub(true, $host, $port, $timeout, $heartbeat);
+        $container = ContainerStub::getHyperfContainer();
+        $container->shouldReceive('make')->with(Socket::class, Mockery::any())->andReturnUsing(function ($_, $args) use ($sock) {
+            return $sock;
+        });
+        $container->shouldReceive('get')->with(StdoutLoggerInterface::class)->andReturnUsing(function () {
+            $logger = Mockery::mock(StdoutLoggerInterface::class);
+            $logger->shouldReceive('error')->once()->with(Mockery::any())->andReturnUsing(function ($message) {
+                $this->assertTrue(Str::contains($message, 'KeepaliveIO heartbeat failed'));
+                $this->assertTrue(Str::contains($message, 'AMQPRuntimeException'));
+                $this->assertTrue(Str::contains($message, 'Socket of keepaliveIO is exhausted.'));
+            });
+
+            return $logger;
+        });
+
+        $io = new KeepaliveIO($host, $port, $timeout, 6, null, true, $heartbeat);
+
+        $io->connect();
+
+        Coroutine::sleep(2);
     }
 }