+++ src/config-aliyun-acm/src/Process/ConfigFetcherProcess.php
@@ -16,6 +16,7 @@ use Hyperf\ConfigAliyunAcm\ClientInterface;
 use Hyperf\ConfigAliyunAcm\PipeMessage;
 use Hyperf\Contract\ConfigInterface;
 use Hyperf\Process\AbstractProcess;
+use Hyperf\Process\ProcessCollector;
 use Psr\Container\ContainerInterface;
 use Swoole\Server;
 
@@ -68,10 +69,21 @@ class ConfigFetcherProcess extends AbstractProcess
             if ($config !== $this->cacheConfig) {
                 $this->cacheConfig = $config;
                 $workerCount = $this->server->setting['worker_num'] + $this->server->setting['task_worker_num'] - 1;
+                $pipeMessage = new PipeMessage($config);
                 for ($workerId = 0; $workerId <= $workerCount; ++$workerId) {
-                    $this->server->sendMessage(new PipeMessage($config), $workerId);
+                    $this->server->sendMessage($pipeMessage, $workerId);
+                }
+
+                $processes = ProcessCollector::all();
+                if ($processes) {
+                    $string = serialize($pipeMessage);
+                    /** @var \Swoole\Process $process */
+                    foreach ($processes as $process) {
+                        $process->exportSocket()->send($string);
+                    }
                 }
             }
+
             sleep($this->config->get('aliyun_acm.interval', 5));
         }
     }