+++ src/json-rpc/src/CoreMiddleware.php
@@ -22,17 +22,14 @@ use Psr\Http\Message\ServerRequestInterface;
 class CoreMiddleware extends \Hyperf\RpcServer\CoreMiddleware
 {
     /**
-     * @var \Hyperf\JsonRpc\ResponseBuilder
+     * @var ResponseBuilder
      */
     protected $responseBuilder;
 
-    public function __construct(ContainerInterface $container, Protocol $protocol, string $serverName)
+    public function __construct(ContainerInterface $container, Protocol $protocol, ResponseBuilder $builder, string $serverName)
     {
         parent::__construct($container, $protocol, $serverName);
-        $this->responseBuilder = make(ResponseBuilder::class, [
-            'dataFormatter' => $protocol->getDataFormatter(),
-            'packer' => $protocol->getPacker(),
-        ]);
+        $this->responseBuilder = $builder;
     }
 
     protected function handleFound(Dispatched $dispatched, ServerRequestInterface $request)
@@ -49,7 +46,7 @@ class CoreMiddleware extends \Hyperf\RpcServer\CoreMiddleware
             $parameters = $this->parseParameters($controller, $action, $request->getParsedBody());
             try {
                 $response = $controllerInstance->{$action}(...$parameters);
-            } catch (\Exception $exception) {
+            } catch (\Throwable $exception) {
                 $response = $this->responseBuilder->buildErrorResponse($request, ResponseBuilder::SERVER_ERROR, $exception);
                 $this->responseBuilder->persistToContext($response);