+++ src/json-rpc/src/DataFormatterFactory.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 namespace Hyperf\JsonRpc;
 
 use Hyperf\Contract\NormalizerInterface;
+use Hyperf\Rpc\Context;
 use Hyperf\Utils\Serializer\SymfonyNormalizer;
 use Psr\Container\ContainerInterface;
 
@@ -22,9 +23,10 @@ class DataFormatterFactory
     {
         /** @var NormalizerInterface $normalizer */
         $normalizer = $container->get(NormalizerInterface::class);
+        $context = $container->get(Context::class);
         if ($normalizer instanceof SymfonyNormalizer) {
-            return new NormalizeDataFormatter($normalizer);
+            return new NormalizeDataFormatter($normalizer, $context);
         }
-        return new DataFormatter();
+        return new DataFormatter($context);
     }
 }