+++ src/json-rpc/src/Pool/RpcConnection.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 namespace Hyperf\JsonRpc\Pool;
 
 use Hyperf\Contract\ConnectionInterface;
+use Hyperf\LoadBalancer\Node;
 use Hyperf\Pool\Connection as BaseConnection;
 use Hyperf\Pool\Exception\ConnectionException;
 use Hyperf\Pool\Pool;
@@ -36,10 +37,9 @@ class RpcConnection extends BaseConnection implements ConnectionInterface
      * @var array
      */
     protected $config = [
-        'host' => 'localhost',
-        'port' => 9501,
+        'node' => null,
         'connect_timeout' => 5.0,
-        'options' => [],
+        'settings' => [],
     ];
 
     public function __construct(ContainerInterface $container, Pool $pool, array $config)
@@ -77,12 +77,18 @@ class RpcConnection extends BaseConnection implements ConnectionInterface
 
     public function reconnect(): bool
     {
-        $host = $this->config['host'];
-        $port = $this->config['port'];
+        if (! $this->config['node'] instanceof \Closure) {
+            throw new ConnectionException('Node of Connection is invalid.');
+        }
+
+        /** @var Node $node */
+        $node = value($this->config['node']);
+        $host = $node->host;
+        $port = $node->port;
         $connectTimeout = $this->config['connect_timeout'];
 
         $client = new SwooleClient(SWOOLE_SOCK_TCP);
-        $client->set($this->config['options'] ?? []);
+        $client->set($this->config['settings'] ?? []);
         $result = $client->connect($host, $port, $connectTimeout);
         if ($result === false && ($client->errCode === 114 || $client->errCode === 115)) {
             // Force close and reconnect to server.
@@ -97,6 +103,7 @@ class RpcConnection extends BaseConnection implements ConnectionInterface
 
     public function close(): bool
     {
+        $this->lastUseTime = 0.0;
         $this->connection->close();
         return true;
     }