+++ src/service-governance/src/ServiceManager.php
@@ -24,7 +24,11 @@ class ServiceManager
      */
     public function register(string $name, string $path, array $metadata): void
     {
-        $this->services[$name][$path] = $metadata;
+        if (isset($metadata['protocol'])) {
+            $this->services[$name][$path][$metadata['protocol']] = $metadata;
+        } else {
+            $this->services[$name][$path]['default'] = $metadata;
+        }
     }
 
     /**